# src/main/java/edu/senac/dto/OlaMundoRequest.java
package edu.senac.dto;

import org.springframework.validation.annotation.Validated;

import com.fasterxml.jackson.annotation.JsonProperty;

@Validated
public class OlaMundoRequest {

	@JsonProperty("nome")
	private String nome = null;

	public String getNome() {
		return nome;
	}

	public void setNome(String nome) {
		this.nome = nome;
	}
	
}