# src/main/java/edu/senac/dto/OlaMundoResponse.java
package edu.senac.dto;

import org.springframework.validation.annotation.Validated;

import com.fasterxml.jackson.annotation.JsonProperty;

@Validated
public class OlaMundoResponse {

	@JsonProperty("mensagem")
	private String mensagem = null;

	public String getMensagem() {
		return mensagem;
	}

	public void setMensagem(String mensagem) {
		this.mensagem = mensagem;
	}

}